libbladeRF, bladeRF-cli: [WIP] First in series of image format changes.
Incomplete and does not build.

This and the following set of changes will simplify the image format
a bit; the metadata fields are flattened into the image to simplify
both image creation and manually viewing/extracting contents from
images.jj

The checksum has been moved to directly after the file format's magic
value (i.e., signature), to make it a bit easier to just extract data
via dd. Similar to the approach used by mkimage for U-Boot headers,
the checksum calculation is performed over the entire image, with the
checksum field zeroed out.